fix potential SIGV on FreeBSD at exit at expense of leaking a little memory
If running under FreeBSD, and perl version < 5.19.6, we must use
PL_use_safe_putenv = 1 in order to avoid a possible SIGV at exit.  This
does cause sentenv to leak a little memory, but seems to be the only way
to avoid the SIGV.  This is only enabled on FreeBSD on perl < 5.19.6.
Newer perls are immune, and I am not currently aware of other platforms
with this issue.